python/requirements.txt
CherryPy==10.2.1
ImageHash==3.4
pillow==3.3.0
